feat: improve message on empty image screen Fixes https://github.com/containers/desktop/issues/56 also move screen to a separate component
