+++ packages/renderer/src/lib/ContainerList.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-import Fa from 'svelte-fa/src/fa.svelte';
-import { faPaste } from '@fortawesome/free-solid-svg-icons';
 import { onMount } from 'svelte';
 import { filtered, searchPattern } from '../stores/containers';
 import { providerInfos } from '../stores/providers';
@@ -10,6 +8,7 @@ import { router } from 'tinro';
 import ContainerDetails from './ContainerDetails.svelte';
 import type { ContainerInfoUI } from './container/ContainerInfoUI';
 import ContainerActions from './container/ContainerActions.svelte';
+import ContainerEmptyScreen from './container/ContainerEmptyScreen.svelte';
 
 let openChoiceModal = false;
 let fromDockerfileModal = false;
@@ -36,11 +35,7 @@ $: {
   }
 }
 
-let copyText;
-
 onMount(async () => {
-  copyText = document.getElementById('noContainerCommandLine') as HTMLElement;
-
   filtered.subscribe(value => {
     containers = value.map((containerInfo: ContainerInfo) => {
       return {
@@ -64,11 +59,6 @@ onMount(async () => {
     }
   });
 });
-
-function copyRunInstructionToClipboard() {
-  const text = copyText?.innerText;
-  navigator.clipboard.writeText(text);
-}
 
 function expandContainerSelection(container: ContainerInfoUI) {
   if (selectedContainer?.name === container.name) {
@@ -348,22 +338,7 @@ function getEngine(containerInfo: ContainerInfo): string {
       {#if selectedContainer}
         <ContainerDetails container="{selectedContainer}" />
       {/if}
-    </div>
-  </div>
-</div>
-<div class="h-full min-w-full flex flex-col" class:hidden="{containers.length > 0}">
-  <div class="pf-c-empty-state h-full">
-    <div class="pf-c-empty-state__content">
-      <i class="fas fa-cubes pf-c-empty-state__icon" aria-hidden="true"></i>
-
-      <h1 class="pf-c-title pf-m-lg">No containers</h1>
-
-      <div class="pf-c-empty-state__body">Run a first container using the following command line:</div>
-      <div class="flex flex-row bg-gray-800 w-full items-center p-2 mt-2">
-        <div id="noContainerCommandLine">podman run redhat/ubi8-micro echo hello world</div>
-        <button title="Copy To Clipboard" class="mr-5" on:click="{() => copyRunInstructionToClipboard()}"
-          ><Fa class="ml-3 h-5 w-5 cursor-pointer rounded-full text-3xl text-sky-800" icon="{faPaste}" /></button>
-      </div>
+      <ContainerEmptyScreen containers="{$filtered}" />
     </div>
   </div>
 </div>
